Development_Phase/requirements_hear.txt
# HeAR SPRSound Training Requirements
# Python 3.8+ required

# Core PyTorch and transformers
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers==4.50.3

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
scipy>=1.10.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Training utilities
tqdm>=4.65.0
python-dotenv>=1.0.0

# HuggingFace
huggingface-hub>=0.19.0

# TensorFlow (optional, for HeAR event detector evaluation)
tensorflow>=2.13.0

# Jupyter (optional, for notebooks)
# jupyter>=1.0.0
# ipywidgets>=8.0.0

# Development tools (optional)
# black>=23.0.0
# flake8>=6.0.0
# pytest>=7.4.0
